refactor(shop): minPrice 타입 Integer 으로 변경

+++ src/main/java/com/example/outsourcing/domain/shop/dto/response/PageShopResponseDto.java
@@ -10,7 +10,7 @@ public class PageShopResponseDto {
     private String address;
     private ShopCategory category;
     private Double averageRating;
-    private Double minPrice;
+    private Integer minPrice;
 
     ShopService shopService;
